+++ packages/fuselage/src/components/CardGroup/index.ts
@@ -1,2 +1,1 @@
 export * from './CardGroup';
-export * from './CardGrid';
